move styling consts to sub-folder, change unit tests
