+++ packages/app/src/components/Dimensions/DimensionItem.js
@@ -10,30 +10,8 @@ import {
     OrgUnitIcon,
     GenericDimension,
 } from '../../icons';
-import { colors } from '../../colors';
 import { setDataTransfer } from '../../dnd';
-
-const style = {
-    wrapper: {
-        display: 'flex',
-        position: 'static',
-    },
-    text: {
-        color: colors.black,
-        cursor: 'pointer',
-        userSelect: 'none',
-        wordBreak: 'break-all',
-        paddingTop: 3,
-        fontSize: 16,
-        maxWidth: 195,
-    },
-    itemContainer: {
-        display: 'flex',
-        minHeight: 24,
-        marginTop: 7,
-        marginBottom: 7,
-    },
-};
+import { styles } from './styles/DimensionItem.style';
 
 const fixedDimensionIcons = {
     dx: <DataIcon />,
@@ -63,7 +41,7 @@ export class DimensionItem extends Component {
         return (
             <span
                 data-dimensionid={this.props.id}
-                style={style.text}
+                style={styles.text}
                 draggable="true"
                 onDragStart={this.onDragStart}
             >
@@ -79,7 +57,7 @@ export class DimensionItem extends Component {
         return (
             <li
                 key={this.props.id}
-                style={style.itemContainer}
+                style={styles.itemContainer}
                 onMouseOver={this.onMouseOver}
                 onMouseLeave={this.onMouseExit}
             >
