+++ packages/app/src/components/Dimensions/DimensionOptions.js
@@ -6,19 +6,7 @@ import MenuItem from '@material-ui/core/MenuItem';
 import DropDown from './DropDown';
 import { MoreHorizontal } from '../../icons';
 import { acAddUiLayoutDimensions } from '../../actions/ui';
-
-const style = {
-    wrapper: {
-        height: 24,
-    },
-    dropDownButton: {
-        border: 'none',
-        background: 'none',
-        outline: 'none',
-        padding: 0,
-        height: 24,
-    },
-};
+import { styles } from './styles/DimensionOptions.style';
 
 const items = [
     {
@@ -37,7 +25,7 @@ const items = [
 
 export const OptionsButton = ({ action }) => {
     return (
-        <button style={style.dropDownButton} onClick={action} tabIndex={1}>
+        <button style={styles.dropDownButton} onClick={action} tabIndex={1}>
             <MoreHorizontal />
         </button>
     );
@@ -82,7 +70,7 @@ export class DimensionOptions extends Component {
         const OptionsButton = this.renderOptionsOnHover();
 
         return (
-            <div style={style.wrapper}>
+            <div style={styles.wrapper}>
                 {OptionsButton}
                 <DropDown
                     id={this.props.id}
