+++ packages/app/src/components/Dimensions/styles/DimensionList.style.js
@@ -1,0 +1,10 @@
+export const styles = {
+    listContainer: {
+        overflow: 'hidden',
+        overflowY: 'scroll',
+        minHeight: 815,
+        minWidth: 250,
+        padding: 0,
+        marginTop: 0,
+    },
+};
