+++ packages/app/src/components/Dimensions/styles/DimensionOptions.style.js
@@ -1,0 +1,12 @@
+export const styles = {
+    wrapper: {
+        height: 24,
+    },
+    dropDownButton: {
+        border: 'none',
+        background: 'none',
+        outline: 'none',
+        padding: 0,
+        height: 24,
+    },
+};
