+++ packages/app/src/components/Dimensions/styles/Dimensions.style.js
@@ -1,0 +1,15 @@
+import { colors } from '../../../colors';
+
+export const styles = {
+    divContainer: {
+        backgroundColor: colors.lightGrey,
+        width: 250,
+        padding: 12,
+    },
+    searchIcon: {
+        color: colors.grey,
+    },
+    textField: {
+        marginBottom: 12,
+    },
+};
