+++ packages/app/src/components/Dimensions/styles/RecommendedIcon.style.js
@@ -1,0 +1,23 @@
+import { colors } from '../../../colors';
+
+export const styles = {
+    toolTip: {
+        color: colors.white,
+        backgroundColor: '#4a4a4a',
+        boxShadow: 'none',
+        width: 150,
+        borderRadius: 3,
+        position: 'relative',
+        top: 5,
+        fontSize: 12,
+        padding: 5,
+    },
+    recommendedIcon: {
+        backgroundColor: colors.accentSecondaryLight,
+        height: 7,
+        width: 7,
+        marginTop: 10,
+        marginLeft: 4,
+        borderRadius: 5,
+    },
+};
